backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
numpy==2.3.2
pandas==2.3.1
pillow==11.3.0
